padronizando entities
