+++ src/main/java/br/com/library/library/entity/AbstractInformational.java
@@ -1,19 +1,19 @@
 package br.com.library.library.entity;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
+import java.io.Serializable;
 
 @Data
 @Entity
 @Table(name = "abstract_informational")
 @NoArgsConstructor
 @AllArgsConstructor
-public class AbstractInformational {
+public class AbstractInformational implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
