+++ src/main/java/br/com/library/library/entity/Book.java
@@ -1,18 +1,14 @@
-package br.com.library.library;
+package br.com.library.library.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 
-@Getter
-@Setter
+@Data
+@Entity
+@Table(name = "book")
 @NoArgsConstructor
 @AllArgsConstructor
 public class Book implements Serializable {
@@ -20,13 +16,10 @@ public class Book implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long isbn;
-
     @NotNull
     private Integer year;
-
     @NotNull
     private Integer volume;
-
     @NotNull
     private Integer edition;
